将 Service 共通代码抽象出来，另作一类
